# arcanium/api/openapi/arcanium.yaml
openapi: 3.1.0
info:
  title: Arcanium API
  version: "1.0.0"
  description: >
    Arcanium's cryptographic control-plane API. Authored from the actual
    Express route handlers (arcanium/api/src/routes/*.js,
    arcanium/api/src/auth/index.js) as source of truth for current
    behavior — a capture-then-enforce exercise (Prompt 22), not a redesign.
    This is the ONE authoritative contract: the Nuxt gateway allowlist
    (arcanium/ui/server/routes/gateway/[...path].ts), the CLI request
    shapes, and docs/api.md's narrative must all be checked against this
    file, not maintained as independent parallel sources of truth.
servers:
  - url: http://localhost:3001
    description: Direct API (dev/scenario scripts)
  - url: http://localhost:3000/gateway
    description: Same-origin UI gateway (browser-facing; allowlisted subset only)
security:
  - sessionCookie: []
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: arc_session
      description: >
        Opaque session id issued by GET /api/v1/auth/callback. When
        ARCANIUM_AUTH_ENABLED is unset, requireSession no-ops and every
        request is treated as an estate-wide operator — see
        docs/security.md.
    serviceAccountBearer:
      type: http
      scheme: bearer
      description: >
        Reserved for a future machine-to-machine identity source
        (prompts/26_external_integration_key_lifecycle.md, not yet
        executed). Not accepted by any route in this version of the API —
        listed here only so the contract does not need a breaking change
        once it is.
  responses:
    Unauthorized:
      description: No session cookie, or the session is expired/invalid.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    BadRequest:
      description: Request failed field validation.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
...
    Error:
      type: object
      properties:
        error:
          description: Direct-API responses use a string; gateway responses use boolean true.
          oneOf:
            - type: string
            - type: boolean
        statusCode: { type: integer }
        statusMessage: { type: string }
        message: { type: string }
        field: { type: string, description: "Direct-API 400 responses name the offending field." }
        reason: { type: string, description: "authorize() denial reason, on 403 responses." }
        action: { type: string, description: "The authorize() action that was denied, on 403 responses." }
      required: [error]

    # ── Suppliers / Applications ─────────────────────────────────────────
...
    Identity:
      type: object
      properties:
        enabled: { type: boolean }
        user: { type: string }
        persona: { type: string }
        namespaces: { type: array, items: { type: string } }
        groups: { type: array, items: { type: string } }
        demoSwitch: { type: boolean }
paths:
...
  /api/v1/auth/callback:
    get:
      summary: OIDC callback — exchanges the code, creates the session, sets arc_session
      security: []
      parameters:
        - { name: code, in: query, schema: { type: string } }
        - { name: state, in: query, schema: { type: string } }
      responses:
        "302":
          description: Redirect to the app. Relayed by a dedicated Nitro route for the same reason as /auth/login.
          headers:
            Location: { schema: { type: string } }
            Set-Cookie: { schema: { type: string } }
        "401": { description: "Invalid state, wrong issuer/audience, or exchange failure" }
  /api/v1/auth/me:
    get:
      summary: Current session identity
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Identity" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /api/v1/auth/demo-persona:
    post:
      summary: Switch the demo session's displayed persona (presentation only — never a real privilege change)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties: { persona: { type: string } }
      responses:
        "200": { description: OK }
        "400": { $ref: "#/components/responses/BadRequest" }

  # ── Applications ──────────────────────────────────────────────────────
...